+++ packages/@webex/internal-plugin-metrics/test/unit/spec/call-diagnostic/call-diagnostic-metrics.ts
@@ -16,7 +16,7 @@ const userAgent = `webex-js-sdk/test-webex-version client=Cantina; (os=${getOSNa
   getOSVersion().split('.')[0]
 })`;
 
-describe.only('internal-plugin-metrics', () => {
+describe('internal-plugin-metrics', () => {
   describe('CallDiagnosticMetrics', () => {
     var now = new Date();
 
@@ -49,7 +49,9 @@ describe.only('internal-plugin-metrics', () => {
           metrics: {
             submitClientMetrics: sinon.stub(),
           },
-          newMetrics: {},
+          newMetrics: {
+            postPreLoginMetric: sinon.stub(),
+          },
           device: {
             userId: 'userId',
             url: 'deviceUrl',
@@ -78,6 +80,11 @@ describe.only('internal-plugin-metrics', () => {
           isUnverifiedGuest: false,
         },
         prepareFetchOptions: sinon.stub().callsFake((opts: any) => ({...opts, foo: 'bar'})),
+        request: sinon.stub().resolves({body: {}}),
+        logger: {
+          log: sinon.stub(),
+          error: sinon.stub(),
+        },
       };
 
       webex.internal.newMetrics.callDiagnosticLatencies = new CallDiagnosticLatencies(
@@ -160,10 +167,10 @@ describe.only('internal-plugin-metrics', () => {
             browser: getBrowserName(),
             browserVersion: getBrowserVersion(),
             clientType: 'TEAMS_CLIENT',
-            "clientVersion": "43.9.0.1234",
-            "localNetworkPrefix": "1.3.4.0",
-            "majorVersion": 43,
-            "minorVersion": 9,
+            clientVersion: '43.9.0.1234',
+            localNetworkPrefix: '1.3.4.0',
+            majorVersion: 43,
+            minorVersion: 9,
             os: getOSNameInternal(),
             osVersion: getOSVersion(),
             subClientType: 'WEB_APP',
@@ -173,7 +180,7 @@ describe.only('internal-plugin-metrics', () => {
           networkType: 'unknown',
           userAgent,
         });
-      })
+      });
     });
 
     describe('#getIdentifiers', () => {
@@ -621,6 +628,28 @@ describe.only('internal-plugin-metrics', () => {
           }
         );
       });
+
+      it('should submit client to diagnostic when no preLoginId provided', () => {
+        const testEvent = {name: 'client.alert.displayed', options: {meetingId: 'meetingId'}};
+        sinon.stub(cd, 'prepareClientEvent').returns(testEvent);
+        sinon.stub(cd, 'submitToCallDiagnostics');
+        sinon.stub(cd, 'submitToCallDiagnosticsPreLogin');
+        //@ts-ignore
+        cd.submitClientEvent(testEvent);
+        assert.calledWith(cd.submitToCallDiagnostics, testEvent);
+        assert.notCalled(cd.submitToCallDiagnosticsPreLogin);
+      });
+
+      it('should submit event to prelogin when preLoginId provided', () => {
+        const testEvent = {name: 'client.alert.displayed', options: {preLoginId: 'preLoginId'}};
+        sinon.stub(cd, 'prepareClientEvent').returns(testEvent);
+        sinon.stub(cd, 'submitToCallDiagnosticsPreLogin');
+        sinon.stub(cd, 'submitToCallDiagnostics');
+        //@ts-ignore
+        cd.submitClientEvent(testEvent);
+        assert.calledWith(cd.submitToCallDiagnosticsPreLogin, testEvent);
+        assert.notCalled(cd.submitToCallDiagnostics);
+      });
     });
 
     it('should send request to call diagnostic batcher', () => {
@@ -752,6 +781,7 @@ describe.only('internal-plugin-metrics', () => {
         );
       });
     });
+
     describe('#getErrorPayloadForClientErrorCode', () => {
       it('it should grab the payload for client error code correctly', () => {
         const res = cd.getErrorPayloadForClientErrorCode({
@@ -779,7 +809,7 @@ describe.only('internal-plugin-metrics', () => {
     });
 
     describe('#generateClientEventErrorPayload', () => {
-      const defaultExpectedRes =  {
+      const defaultExpectedRes = {
         category: 'expected',
         errorDescription: 'StartRecordingFailed',
         fatal: true,
@@ -788,7 +818,7 @@ describe.only('internal-plugin-metrics', () => {
         errorCode: 4029,
         serviceErrorCode: 2409005,
       };
-      
+
       const checkNameError = (payload: any, isExpectedToBeCalled: boolean) => {
         const res = cd.generateClientEventErrorPayload(payload);
         const expectedResult = {
@@ -807,7 +837,7 @@ describe.only('internal-plugin-metrics', () => {
         } else {
           assert.notDeepEqual(res, expectedResult);
         }
-      }
+      };
 
       it('should generate media event error payload if rawError has a media error name', () => {
         checkNameError({name: 'PermissionDeniedError'}, true);
@@ -820,13 +850,13 @@ describe.only('internal-plugin-metrics', () => {
       const checkCodeError = (payload: any, expetedRes: any) => {
         const res = cd.generateClientEventErrorPayload(payload);
         assert.deepEqual(res, expetedRes);
-      }
+      };
       it('should generate event error payload correctly', () => {
-        checkCodeError({body: {errorCode: 2409005}}, defaultExpectedRes)
+        checkCodeError({body: {errorCode: 2409005}}, defaultExpectedRes);
       });
 
       it('should generate event error payload correctly if rawError has body.code', () => {
-        checkCodeError({body: {code: 2409005}}, defaultExpectedRes)
+        checkCodeError({body: {code: 2409005}}, defaultExpectedRes);
       });
 
       it('should generate event error payload correctly if rawError has body.reason.reasonCode', () => {
@@ -854,7 +884,7 @@ describe.only('internal-plugin-metrics', () => {
         } else {
           assert.notDeepEqual(res, expectedResult);
         }
-      }
+      };
 
       it('should return default new locus event error payload correctly if locus error is recognized', () => {
         checkLocusError({body: {errorCode: 2400000}}, true);
@@ -881,7 +911,7 @@ describe.only('internal-plugin-metrics', () => {
         } else {
           assert.notDeepEqual(res, expectedResult);
         }
-      }
+      };
 
       it('should return default meeting info lookup error payload if data.meetingInfo was found on error body', () => {
         checkMeetingInfoError({body: {data: {meetingInfo: 'something'}}}, true);
@@ -890,7 +920,6 @@ describe.only('internal-plugin-metrics', () => {
       it('should return default meeting info lookup error payload if body.url contains wbxappapi', () => {
         checkMeetingInfoError({body: {url: '1234567-wbxappapiabcdefg'}}, true);
       });
-
 
       it('should not return default meeting info lookup error payload if body.url does not contain wbxappapi and data.meetingInfo was not found on error body', () => {
         checkMeetingInfoError({body: {data: '1234567-wbxappapiabcdefg'}}, false);
@@ -929,13 +958,15 @@ describe.only('internal-plugin-metrics', () => {
       it('returns false if converged architecture is not enabled', () => {
         fakeMeeting.meetingInfo = {enableConvergedArchitecture: false};
         assert.deepEqual(cd.getIsConvergedArchitectureEnabled({meetingId: fakeMeeting.id}), false);
-
       });
       it('returns undefined if converged architecture is not defined', () => {
         fakeMeeting.meetingInfo = {};
-        assert.deepEqual(cd.getIsConvergedArchitectureEnabled({meetingId: fakeMeeting.id}), undefined);
+        assert.deepEqual(
+          cd.getIsConvergedArchitectureEnabled({meetingId: fakeMeeting.id}),
+          undefined
+        );
       });
-    })
+    });
 
     describe('#buildClientEventFetchRequestOptions', () => {
       it('returns expected options', async () => {
@@ -1009,6 +1040,52 @@ describe.only('internal-plugin-metrics', () => {
           resource: 'clientmetrics',
           service: 'metrics',
         });
+      });
+    });
+
+    describe('#submitToCallDiagnosticsPreLogin', async () => {
+      it('should call webex.request with expected options', async () => {
+        sinon.spy(Utils, 'prepareDiagnosticMetricItem');
+        await cd.submitToCallDiagnosticsPreLogin(
+          {
+            //@ts-ignore
+            event: {name: 'client.alert.displayed', canProceed: true},
+            //@ts-ignore
+            originTime: {triggered: 'now'},
+          },
+          'my-id'
+        );
+
+        assert.calledWith(Utils.prepareDiagnosticMetricItem, webex, {
+          eventPayload: {
+            event: {name: 'client.alert.displayed', canProceed: true},
+            originTime: {triggered: 'now', sent: now.toISOString()},
+            origin: {buildType: 'test', networkType: 'unknown'},
+          },
+          type: ['diagnostic-event'],
+        });
+
+        assert.calledWith(
+          webex.internal.newMetrics.postPreLoginMetric,
+          {
+            eventPayload: {
+              event: {
+                name: 'client.alert.displayed',
+                canProceed: true,
+              },
+              originTime: {
+                sent: now.toISOString(),
+                triggered: 'now',
+              },
+              origin: {
+                buildType: 'test',
+                networkType: 'unknown',
+              },
+            },
+            type: ['diagnostic-event'],
+          },
+          'my-id'
+        );
       });
     });
   });